src/test/javascript/spec/app/entities/formation-sfr/formation-sfr.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { FormationSFRService } from 'app/entities/formation-sfr/formation-sfr.service';
import { IFormationSFR, FormationSFR } from 'app/shared/model/formation-sfr.model';

describe('Service Tests', () => {
  describe('FormationSFR Service', () => {
    let injector: TestBed;
    let service: FormationSFRService;
    let httpMock: HttpTestingController;
    let elemDefault: IFormationSFR;
    let expectedResult: IFormationSFR | IFormationSFR[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(FormationSFRService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new FormationSFR(0, 0, 0, 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a FormationSFR', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new FormationSFR()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a FormationSFR', () => {
        const returnedFromService = Object.assign(
          {
            annee: 1,
            nbrAgentForme: 1,
            themeFormation: 'BBBBBB',
            dureeTotFormation: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of FormationSFR', () => {
        const returnedFromService = Object.assign(
          {
            annee: 1,
            nbrAgentForme: 1,
            themeFormation: 'BBBBBB',
            dureeTotFormation: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a FormationSFR', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
